1 Версия ТЗ11

+++ src/main/java/ru/yandex/practicum/filmorate/storage/FilmStorage.java
@@ -12,7 +12,7 @@ public interface FilmStorage {
 
     Film updateFIlm(Film film);
 
-    void deleteFilm(long id);
+    void deleteFilm(int id);
 
-    Film getFilmById(long id);
+    Film getFilmById(int id);
 }
